requirements.txt
bluepy==1.3.0
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.1
plotly==5.14.1
pyudev==0.24.1
pyzed==1.3.0
scikit_learn==1.2.2
streamlit==1.22.0
streamlit_toggle_switch==1.0.2
